+++ CHANGELOG.rst
@@ -14,7 +14,8 @@ may want to subscribe to `GitHub's tag feed
 
 * NEW ikhal's event editor now allows better editing of recurrence rules,
   including INTERVALs, end dates, and more
-
+* NEW ikhal will now check if any configured vdir has been updated, and, if
+  applicable, refresh its UI to reflect the latest changes 
 
 0.9.3
 =====